+++ src/osdmavlink.c
@@ -87,7 +87,7 @@ void parseMavlink(void)
                         apm_mav_system    = msg.sysid;
                         apm_mav_component = msg.compid;
                         apm_mav_type      = mavlink_msg_heartbeat_get_type(&msg);
-                        osd_mode = (uint8_t)mavlink_msg_heartbeat_get_custom_mode(&msg);
+                        osd_mode = mavlink_msg_heartbeat_get_custom_mode(&msg);
                         base_mode = mavlink_msg_heartbeat_get_base_mode(&msg);
 
                         last_motor_armed = motor_armed;
